Change all primary key types int -> long (4byte -> 8byte)
많은 데이터를 처리할 수 있는 타입으로 변환

+++ backend/app/src/main/java/com/vms/app/entity/Appointment.java
@@ -30,7 +30,7 @@ public class Appointment {
   @Id()
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(length = 45, name = "appointmentID")
-  private int appointmentID;
+  private long appointmentID;
 
   @Column(length = 100)
   private String date; // 방문 날짜
