+++ backend/app/src/main/java/com/vms/app/entity/Notice.java
@@ -23,7 +23,7 @@ public class Notice {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column()
-  private int noticeID;
+  private long noticeID;
 
   @Column(length = 50, nullable = false)
   private String title;
